Add SerializedView.
It contains common code used by view which return serialized data.
